refactor member data organization
